+++ tests/payment/Payment - Gift card - Full order value - vtexgame1.test.js
@@ -1,0 +1,3 @@
+import test from './models/Payment - Gift card - Full order value.model.js'
+
+test('vtexgame1')